+++ include/utils/filesystem.h
@@ -1,0 +1,18 @@
+/*!
+ * \file include/utils/filesystem.h
+ * \brief file-system utilities
+ *
+ * \author xythobuz
+ */
+
+#ifndef _UTILS_FILESYSTEM_H_
+#define _UTILS_FILESYSTEM_H_
+
+#include <string>
+
+std::string getCurrentWorkingDirectory();
+
+std::string getHomeDirectory();
+
+#endif
+