+++ src/utils/File.cpp
@@ -1,0 +1,19 @@
+/*!
+ * \file src/utils/File.cpp
+ * \brief Recursive file-system walking utilities
+ *
+ * \author xythobuz
+ */
+
+#include "global.h"
+#include "utils/File.h"
+#include "utils/Folder.h"
+
+File::File(std::string file) {
+
+}
+
+File::~File() {
+
+}
+