+++ src/utils/Folder.cpp
@@ -1,0 +1,19 @@
+/*!
+ * \file src/utils/Folder.cpp
+ * \brief Recursive file-system walking utilities
+ *
+ * \author xythobuz
+ */
+
+#include "global.h"
+#include "utils/File.h"
+#include "utils/Folder.h"
+
+Folder::Folder(std::string folder) {
+
+}
+
+Folder::~Folder() {
+
+}
+